This should fix the fps limiter if you play longer than 2^32 nanoseconds.

+++ code/branches/presentation2/src/libraries/core/Game.cc
@@ -313,7 +313,7 @@ namespace orxonox
 
     void Game::updateFPSLimiter()
     {
-        uint32_t nextTime = gameClock_->getMicroseconds() - excessSleepTime_ + static_cast<uint32_t>(1000000.0f / fpsLimit_);
+        uint64_t nextTime = gameClock_->getMicroseconds() - excessSleepTime_ + static_cast<uint32_t>(1000000.0f / fpsLimit_);
         uint64_t currentRealTime = gameClock_->getRealMicroseconds();
         while (currentRealTime < nextTime - minimumSleepTime_)
         {
